services/ingestor/requirements.txt
requests
xmltodict
python-dotenv
pydantic
azure-storage-blob
